requirements.txt
Flask==3.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
supabase==2.16.0
flask-cors==4.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
requests==2.31.0

# Development dependencies
black==23.11.0
flake8==6.1.0
isort==5.12.0
safety==2.3.5
bandit==1.7.5 